Define RGG and Mesh subclasses of Network
